ARM: dts: bcm2711-rpi-400: Don't delete the ACT LED
The Pi 400 only has one user-accessible LED (green), which is intended
to be used as a power indication. Therefore there is no activity LED.
However, a user may wish to remap the activity LED functionality to a
GPIO in the 40-way header, so preserve the led-act node in a disabled
state.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>